201403762 - Correccion de pruebas para asegurar el code coverage

+++ E-commerce/utilities/OperacionesCategoria.php
@@ -8,37 +8,27 @@ use app\models\Producto;
 
 class OperacionesCategoria{
   public static function Nueva_Categoria($nombre_categoria, $codigo){
-    try{
-          $categoria = new Categoria;
-          $categoria->nombre = $nombre_categoria;
-          $categoria->codigo = $codigo;
-          $categoria->save();
-          return true;
-    } catch(Exception $e){
-
-    } // catch
+      if($nombre_categoria!=null&&$codigo!=null){
+        $categoria = new Categoria;
+        $categoria->nombre = $nombre_categoria;
+        $categoria->codigo = $codigo;
+        $categoria->save();
+        return true;
+      }
       return false;
   }
-  public static function Quitar_Categoria($codigo_producto){
-    try{
-      $producto_row = Categoria::find()->where('codigo = :codigo', [':codigo' => $codigo_producto])->one();
-      $producto_row->delete();
-      return true;
-    } catch(Exception $e){
 
-    } // catch
+  public static function Quitar_Categoria($codigo_producto){
+      $producto_row = Categoria::find()->where('codigo = :codigo', [':codigo' => $codigo_producto])->one();
+      if($producto_row!=null){
+        $producto_row->delete();
+        return true;
+      }
       return false;
   }
 
   public static function get_categorias_disponibles(){
       $query_string= "SELECT * from categoria";
-      $connection = Yii::$app->getDb();
-      $command = $connection->createCommand($query_string);
-      $result = $command->queryAll();
-      return $result;
-  }
-  public static function get_productos_comprados(){
-      $query_string= "SELECT * from Producto where estado = 2;";
       $connection = Yii::$app->getDb();
       $command = $connection->createCommand($query_string);
       $result = $command->queryAll();
